Added in everything from profiles page
